fix:Bouncing window multi-layer problem
