+++ src/components/relyList/RelyList.jsx
@@ -1,12 +1,7 @@
 import React, { useContext } from 'react'
 import ButtonUI from '../../UIComponents/buttons/ButtonUI'
 import ReplyListItem from '../../UIComponents/listItems/ReplyListItem'
-import ReplyModal from '../replyModal/ReplyModal'
-import {
-  ShowReplyModel,
-  ReplyModelIsShow,
-  HideModel,
-} from '../../contexts/modalControlContext/ModalControlContext'
+import { ShowReplyModel } from '../../contexts/modalControlContext/ModalControlContext'
 
 import style from './ReplyList.module.scss'
 
@@ -19,8 +14,6 @@ export default function ReplyList({
   likeCount,
 }) {
   const handleShowReplyModel = useContext(ShowReplyModel)
-  const handleHideModel = useContext(HideModel)
-  const replyModelIsShow = useContext(ReplyModelIsShow)
   return (
     <div className={style.viewport}>
       <div className={style.tweet__container}>
@@ -48,8 +41,8 @@ export default function ReplyList({
           <div className={style.tweet__text}>{tweet}</div>
           <div className={style.createAt}>
             <div className={style.createAt__time}>{time}</div>
-            {/* <div className={style.spot} />
-            <div className={style.createAt__data}>{data}</div> */}
+            {/* {/* <div className={style.spot} /> */}
+            {/* <div className={style.createAt__data}>{data}</div> */}
           </div>
           <div className={style.counts}>
             <div className={style.replyCount}>
@@ -66,7 +59,6 @@ export default function ReplyList({
           <div onClick={handleShowReplyModel} className={style.icon__reply}>
             <img className={style.icon__reply__img} />
           </div>
-          {replyModelIsShow && <ReplyModal handleHideModel={handleHideModel} />}
           <div className={style.icon__like}>
             <img className={style.icon__like__img} />
           </div>
