+++ src/components/replyModal/ReplyModal.jsx
@@ -22,74 +22,85 @@ export default function ReplyModal({ handleHideModel }) {
   const handleReplyChange = (e) => {
     setReply(e.target.value)
   }
+
   const portalElement = document.getElementById('modal-root')
   return (
     <>
       {ReactDOM.createPortal(
         <>
-        <div className={style.view__container}>
-          <Modal onHideModel={handleHideModel} buttonText="回覆" title="推文">
-            <div className={style.flexbox}>
-              <div className={style.other__user__container}>
-                <div className={style.other__avatar__container}>
-                  <div className={style.other__avatar}>
+          <div className={style.view__container}>
+            <Modal onHideModel={handleHideModel} buttonText="回覆" title="推文">
+              <div className={style.flexbox}>
+                <div className={style.other__user__container}>
+                  <div className={style.other__avatar__container}>
+                    <div className={style.other__avatar}>
+                      <img
+                        className={style.other__avatar__img}
+                        src={avatar}
+                        alt="Avatar"
+                      />
+                    </div>
+                    <div className={style.straight__line} />
+                  </div>
+
+                  <div className={style.tweet__container}>
+                    <div className={style.tweet__owner}>
+                      <div className={style.other__user__name}>
+                        {DUMMYDATA.name}
+                      </div>
+                      <div className={style.tweetby}>
+                        <div className={style.other__user__account}>
+                          {`@${DUMMYDATA.account}`}
+                        </div>
+                        <div className={style.spot} />
+                        <div className={style.tweet__time}>
+                          {DUMMYDATA.time}
+                        </div>
+                      </div>
+                    </div>
+
+                    <div className={style.reply__for__mobile}>
+                      <p>回覆</p>
+                      <p className={style.account}>
+                        {' '}
+                        {`@${DUMMYDATA.account}`}
+                      </p>
+                    </div>
+
+                    <div className={style.reply__context}>
+                      {DUMMYDATA.tweet}
+                    </div>
+
+                    <div className={style.reply__for}>
+                      <p>回覆</p>
+                      <p className={style.account}>
+                        {' '}
+                        {`@${DUMMYDATA.account}`}
+                      </p>
+                    </div>
+                  </div>
+                </div>
+
+                <div className={style.modal__main}>
+                  <div className={style.avatar}>
                     <img
-                      className={style.other__avatar__img}
+                      className={style.avatar__img}
                       src={avatar}
                       alt="Avatar"
                     />
                   </div>
-                  <div className={style.straight__line} />
-                </div>
-
-                <div className={style.tweet__container}>
-                  <div className={style.tweet__owner}>
-                    <div className={style.other__user__name}>
-                      {DUMMYDATA.name}
-                    </div>
-                    <div className={style.tweetby}>
-                      <div className={style.other__user__account}>
-                        {`@${DUMMYDATA.account}`}
-                      </div>
-                      <div className={style.spot} />
-                      <div className={style.tweet__time}>{DUMMYDATA.time}</div>
-                    </div>
-                  </div>
-
-                  <div className={style.reply__for__mobile}>
-                    <p>回覆</p>
-                    <p className={style.account}> {`@${DUMMYDATA.account}`}</p>
-                  </div>
-
-                  <div className={style.reply__context}>{DUMMYDATA.tweet}</div>
-
-                  <div className={style.reply__for}>
-                    <p>回覆</p>
-                    <p className={style.account}> {`@${DUMMYDATA.account}`}</p>
+                  <div className={style.input__container}>
+                    <Textarea
+                      textareaPlaceHolder="推你的回覆"
+                      textareaValue={reply}
+                      onChange={handleReplyChange}
+                    />
                   </div>
                 </div>
               </div>
-
-              <div className={style.modal__main}>
-                <div className={style.avatar}>
-                  <img
-                    className={style.avatar__img}
-                    src={avatar}
-                    alt="Avatar"
-                  />
-                </div>
-                <div className={style.input__container}>
-                  <Textarea
-                    textareaPlaceHolder="推你的回覆"
-                    textareaValue={reply}
-                    onChange={handleReplyChange}
-                  />
-                </div>
-              </div>
-            </div>
-          </Modal>
-        </div>
-        <BackDrop onHideModel={handleHideModel} />
+            </Modal>
+          </div>
+          <BackDrop onHideModel={handleHideModel} />
         </>,
         portalElement
       )}
