+++ src/pages/CurrentUserTweet.jsx
@@ -1,8 +1,12 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useContext } from 'react'
 import { useNavigate } from 'react-router-dom'
 import CurrentUser from '../components/currentUser/CurrentUser'
 import TweetListItem from '../UIComponents/listItems/TweetListItem'
-
+import ReplyModal from '../components/replyModal/ReplyModal'
+import {
+  ReplyModelIsShow,
+  HideModel,
+} from '../contexts/modalControlContext/ModalControlContext'
 import userApi from '../API/userApi'
 import { Alert } from '../utils/helpers'
 
@@ -11,7 +15,8 @@ import style from './CurrentUserTweet.scss'
 export default function CurrentUserTweet() {
   const [currentUser, setCurrentUser] = useState([])
   const navigate = useNavigate()
-
+  const handleHideModel = useContext(HideModel)
+  const replyModelIsShow = useContext(ReplyModelIsShow)
   useEffect(() => {
     userApi
       .getCurrentUser()
@@ -34,6 +39,7 @@ export default function CurrentUserTweet() {
 
   return (
     <div className={style.userTweet__container}>
+      {replyModelIsShow && <ReplyModal handleHideModel={handleHideModel} />}
       <CurrentUser
         coverImg={currentUser.cover}
         name={currentUser.name}
