+++ src/pages/Home.jsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
 import MainTweet from '../components/mainTweet/MainTweet'
-
 import tweetApi from '../API/tweetApi'
 import userApi from '../API/userApi'
 import { Alert } from '../utils/helpers'
